LLM/requirements.txt
fastapi
uvicorn
langchain
langchain-community
langchain-text-splitters
langchain-huggingface
langchain-ollama
faiss-cpu
pypdf